Improve record class detection (#428)
Also generate a method mapping for the record entry, does not support the case of the method having a diffrent name from the field.

If there is a component this is remapped via the field mapping.

(cherry-picked from commit 2eb7d2b83a397d4d276c882107fc50ddd7ce866a)